JPA/spring-data-jpa/src/main/kotlin/io/beaniejoy/springdatajpa/repository/MemberJpaRepository.kt
package io.beaniejoy.springdatajpa.repository

import io.beaniejoy.springdatajpa.entity.Member
import jakarta.persistence.EntityManager
import jakarta.persistence.PersistenceContext
import org.springframework.stereotype.Repository

@Repository
class MemberJpaRepository(
    @PersistenceContext
    private val em: EntityManager
) {
    fun save(member: Member): Member {
        em.persist(member)
        return member
    }

    fun find(id: Long): Member? {
        return em.find(Member::class.java, id)
    }
}
